convert userTags composable to store
